Update README.md
There is no spelling or grammar errors in the original readme file, but the contents are not organized. My revision added layout to the readme file.  'Pseudo Code' has been restructured to make it a clearer flow of commands and easy to understand. 'Test' method has also been highlighted for the ease of use.